a) Moving src\SupportingLibraries -> support; b) introducing $(platformtoolset) variable to ensure the same location of untracked GDAL_SDK files across branches; c) adding SupportLibs.sln with spatial index and cqlib. d) renaming MapWinGIS2010.sln -> MapWinGIS.sln e) adding .gitignore;

+++ src/COM classes/OgrLayer.cpp
@@ -677,7 +677,7 @@ STDMETHODIMP COgrLayer::HasLocalChanges(VARIANT_BOOL* retVal)
 		CSingleLock lock(&_loader.ProviderLock);
 		if (_dynamicLoading) lock.Lock();
 
-		int featureCount = _layer->GetFeatureCount();
+		int featureCount = static_cast<int>(_layer->GetFeatureCount());
 		if (numShapes != featureCount)	   // i.e. deleted or inserted feature
 		{
 			*retVal = VARIANT_TRUE;
@@ -780,7 +780,7 @@ STDMETHODIMP COgrLayer::get_FeatureCount(VARIANT_BOOL forceLoading, int* retVal)
 	CSingleLock lock(&_loader.ProviderLock);
 	if (_dynamicLoading) lock.Lock();
 	if (_featureCount == -1 || forceLoading) {
-		_featureCount = _layer->GetFeatureCount(forceLoading == VARIANT_TRUE);
+		_featureCount = static_cast<int>(_layer->GetFeatureCount(forceLoading == VARIANT_TRUE));
 	}
 	*retVal = _featureCount;
 	return S_OK;
@@ -1354,14 +1354,14 @@ void COgrLayer::GetFieldValues(OGRFieldType fieldType, BSTR& fieldName, vector<V
 		OGRLayer* layer = _dataset->ExecuteSQL(OgrHelper::String2OgrString(sql), NULL, NULL);
 		if (layer)
 		{
-			OgrHelper::GetFieldValues(layer, _layer->GetFeatureCount(), fieldType, values, _globalCallback);
+			OgrHelper::GetFieldValues(layer, static_cast<int>(_layer->GetFeatureCount()), fieldType, values, _globalCallback);
 			_dataset->ReleaseResultSet(layer);
 		}
 	}
 	else
 	{
 		// the column can be a computed one, so have to load feature as a whole
-		OgrHelper::GetFieldValues(_layer, _layer->GetFeatureCount(), fieldType, values, _globalCallback);
+		OgrHelper::GetFieldValues(_layer, static_cast<int>(_layer->GetFeatureCount()), fieldType, values, _globalCallback);
 	}
 }
